+++ CKPromiseTests/CKPromiseTestsBase.m
@@ -13,17 +13,4 @@
     promise = [CKPromise promise];
 }
 
-- (NSManagedObjectContext*)managedContext {
-    NSManagedObjectModel *mom = [NSManagedObjectModel mergedModelFromBundles:@[[NSBundle bundleForClass:[self class]]]];
-    NSPersistentStoreCoordinator *psc = [[NSPersistentStoreCoordinator alloc] initWithManagedObjectModel:mom];
-    [psc addPersistentStoreWithType:NSInMemoryStoreType
-                      configuration:nil
-                                URL:nil
-                            options:nil
-                              error:nil];
-    NSManagedObjectContext *ctx = [[NSManagedObjectContext alloc] initWithConcurrencyType:NSPrivateQueueConcurrencyType];
-    ctx.persistentStoreCoordinator = psc;
-    return ctx;
-}
-
 @end