+++ CKSwiftPromise/CKSwiftPromise.h
@@ -1,0 +1,19 @@
+//
+//  CKSwiftPromise.h
+//  CKSwiftPromise
+//
+//  Created by Cristian Kocza on 02/11/15.
+//  Copyright © 2015 Cristik. All rights reserved.
+//
+
+#import "CKPromise.h"
+
+//! Project version number for CKSwiftPromise.
+FOUNDATION_EXPORT double CKSwiftPromiseVersionNumber;
+
+//! Project version string for CKSwiftPromise.
+FOUNDATION_EXPORT const unsigned char CKSwiftPromiseVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <CKSwiftPromise/PublicHeader.h>
+
+